+++ src/EmporiqaIntegration.php
@@ -10,7 +10,7 @@ use Shopware\Core\Framework\Plugin\Context\UninstallContext;
 
 class EmporiqaIntegration extends Plugin
 {
-    public const PLUGIN_VERSION = '1.2.1';
+    public const PLUGIN_VERSION = '1.2.2';
 
     public function uninstall(UninstallContext $uninstallContext): void
     {